Conflicts:
	include/Algo/Render/GL2/explodeVolumeRender.hpp

+++ src/Utils/GLSLShader.cpp
@@ -1079,7 +1079,7 @@ void GLSLShader::enableVertexAttribs(unsigned int stride, unsigned int begin) co
 	{
 		glBindBuffer(GL_ARRAY_BUFFER, it->vbo_ptr->id());
 		glEnableVertexAttribArray(it->va_id);
-		glVertexAttribPointer(it->va_id, it->vbo_ptr->dataSize(), GL_FLOAT, false, stride, (const GLvoid*)begin);
+		glVertexAttribPointer(it->va_id, it->vbo_ptr->dataSize(), GL_FLOAT, false, stride, (const GLvoid*)((unsigned long)(begin)));
 	}
 //	this->unbind();
 }
